refactor(plugin/tests): extract shared helpers to testhelpers_test.go
Move writePluginManifestScript, writeExitScript, and createScript out of
validation_test.go into a dedicated testhelpers_test.go file to make their
shared nature explicit and improve discoverability.
